+++ portal-ui/package.json
@@ -36,6 +36,7 @@
     "chart.js": "^2.9.3",
     "codemirror": "^5.52.2",
     "history": "^4.10.1",
+    "kbar": "^0.1.0-beta.27",
     "local-storage-fallback": "^4.1.1",
     "lodash": "^4.17.21",
     "minio": "^7.0.26",
