+++ portal-ui/src/icons/AzureTierIconXs.tsx
@@ -31,7 +31,6 @@ const AzureTierIconXs = (props: SVGProps<SVGSVGElement>) => {
             id="path21"
             d="M447.781,487.513l5.188-.917.049-.011-2.668-3.173c-1.467-1.746-2.668-3.181-2.668-3.188s2.756-7.6,2.771-7.63c.006-.009,1.881,3.229,4.545,7.847l4.572,7.923.035.062-8.479,0-8.48,0S447.781,487.513,447.781,487.513Zm-10.178-.969s1.257-2.187,2.794-4.85l2.794-4.842,3.257-2.733c1.792-1.5,3.261-2.735,3.266-2.737a.672.672,0,0,1-.052.132c-.035.074-1.627,3.487-3.535,7.583l-3.472,7.448-2.525,0C438.739,486.551,437.6,486.55,437.6,486.544Z"
             transform="translate(-437.603 -471.382)"
-            fill="#8399ab"
           />
         </g>
       </g>
