+++ portal-ui/src/icons/SidebarMenus/BucketsMenuIcon.tsx
@@ -32,7 +32,6 @@ const BucketsMenuIcon = (props: SVGProps<SVGSVGElement>) => (
           data-name="Rectángulo 928"
           width="15.957"
           height="15.928"
-          fill="#8399ab"
         />
       </clipPath>
     </defs>
@@ -61,7 +60,6 @@ const BucketsMenuIcon = (props: SVGProps<SVGSVGElement>) => (
               data-name="Trazado 7002"
               d="M15.619.545A1.341,1.341,0,0,0,14.553,0H1.386A1.34,1.34,0,0,0,.32.545a1.606,1.606,0,0,0-.3,1.242c.325,1.888,1.009,5.869,1.557,9.045v.006c.277,1.616.519,3.023.661,3.84A1.422,1.422,0,0,0,3.6,15.911h8.733A1.423,1.423,0,0,0,13.7,14.679l.659-3.836,0-.023.893-5.2,0-.015.658-3.821a1.6,1.6,0,0,0-.3-1.242M13.187,11.3l-10.426,0-.2-1.189H13.383Zm.89-5.216-12.221,0L1.651,4.9H14.273Z"
               transform="translate(0.061 -0.072)"
-              fill="#8399ab"
             />
           </g>
         </g>
