+++ portal-ui/src/icons/SidebarMenus/InspectMenuIcon.tsx
@@ -29,7 +29,6 @@ const InspectMenuIcon = (props: SVGProps<SVGSVGElement>) => (
       id="InspectIcon"
       d="M-2191.428,31a1.876,1.876,0,0,1-1.715-2V27.5h1.285V29a.47.47,0,0,0,.429.5h6.857a.47.47,0,0,0,.428-.5V27.5h1.286V29a1.877,1.877,0,0,1-1.715,2ZM-2194,26V24h12v2Zm2.142-3.5h-1.284V21a1.876,1.876,0,0,1,1.715-2h6.857a1.876,1.876,0,0,1,1.715,2v1.5h-1.286V21a.469.469,0,0,0-.428-.5h-6.857a.469.469,0,0,0-.429.5v1.5h0Z"
       transform="translate(2194 -19)"
-      fill="#fff"
     />
   </svg>
 );
