+++ portal-ui/src/icons/SidebarMenus/LogsMenuIcon.tsx
@@ -33,7 +33,6 @@ const LogsMenuIcon = (props: SVGProps<SVGSVGElement>) => (
           width="12"
           height="12"
           transform="translate(0 0)"
-          fill="#fff"
         />
       </clipPath>
     </defs>
@@ -49,14 +48,12 @@ const LogsMenuIcon = (props: SVGProps<SVGSVGElement>) => (
           data-name="Trazado 7070"
           d="M.1,86.274v7.138a.806.806,0,0,0,.805.8H11.273a.806.806,0,0,0,.805-.8V86.274Zm4.482,1.274v.764a.324.324,0,0,1-.318.331H1.358a.325.325,0,0,1-.319-.331v-.764a.325.325,0,0,1,.319-.33H4.264a.324.324,0,0,1,.318.33Z"
           transform="translate(-0.135 -82.221)"
-          fill="#fff"
         />
         <path
           id="Trazado_7071"
           data-name="Trazado 7071"
           d="M11.273.1H.905A.806.806,0,0,0,.1.906v2.34H12.078V.906A.806.806,0,0,0,11.273.1"
           transform="translate(-0.135 -0.084)"
-          fill="#fff"
         />
       </g>
     </g>
