+++ portal-ui/src/icons/SidebarMenus/MenuCollapsedIcon.tsx
@@ -46,7 +46,6 @@ const MenuCollapsedIcon = (props: SVGProps<SVGSVGElement>) => (
             data-name="Trazado 6842"
             d="M.422,6.661a.433.433,0,0,1-.3-.117.37.37,0,0,1,0-.557L2.983,3.335.126.675a.37.37,0,0,1,0-.557.443.443,0,0,1,.6,0L3.889,3.052a.373.373,0,0,1,.126.274.344.344,0,0,1-.126.274L.727,6.533a.443.443,0,0,1-.306.127Z"
             transform="translate(0 0)"
-            fill="#fff"
           />
         </g>
         <rect
