+++ portal-ui/src/icons/SidebarMenus/MenuExpandedIcon.tsx
@@ -44,7 +44,6 @@ const MenuCollapsedIcon = (props: SVGProps<SVGSVGElement>) => (
           data-name="Trazado 6842"
           d="M.422,0a.433.433,0,0,0-.3.117.37.37,0,0,0,0,.557L2.983,3.325.126,5.986a.37.37,0,0,0,0,.557.443.443,0,0,0,.6,0L3.889,3.609a.373.373,0,0,0,.126-.274.344.344,0,0,0-.126-.274L.727.127A.443.443,0,0,0,.422,0Z"
           transform="translate(0 0)"
-          fill="#8399ab"
         />
       </g>
       <rect
