+++ portal-ui/src/icons/SidebarMenus/ProfileMenuIcon.tsx
@@ -32,7 +32,6 @@ const ProfileMenuIcon = (props: SVGProps<SVGSVGElement>) => (
           data-name="Rectángulo 1599"
           width="12"
           height="10.456"
-          fill="#fff"
         />
       </clipPath>
     </defs>
@@ -46,21 +45,18 @@ const ProfileMenuIcon = (props: SVGProps<SVGSVGElement>) => (
         data-name="Trazado 7122"
         d="M33.036,1.016H43.058L43.3.207A.161.161,0,0,0,43.145,0h-10.2a.161.161,0,0,0-.154.207Z"
         transform="translate(-32.063)"
-        fill="#fff"
       />
       <path
         id="Trazado_7123"
         data-name="Trazado 7123"
         d="M11.551,67.822H.449A.449.449,0,0,0,0,68.333l.644,4.659a.451.451,0,0,0,.018.078H11.334a.451.451,0,0,0,.018-.078L12,68.333a.449.449,0,0,0-.445-.511"
         transform="translate(0 -66.323)"
-        fill="#fff"
       />
       <path
         id="Trazado_7124"
         data-name="Trazado 7124"
         d="M16.471,328.2H5.652a.476.476,0,0,0-.452.624l.845,2.576H16.078l.845-2.576a.476.476,0,0,0-.452-.624"
         transform="translate(-5.062 -320.942)"
-        fill="#fff"
       />
     </g>
   </svg>
