+++ portal-ui/src/icons/SidebarMenus/UsersMenuIcon.tsx
@@ -32,7 +32,6 @@ const UsersMenuIcon = (props: SVGProps<SVGSVGElement>) => (
           data-name="Rectángulo 991"
           width="9.008"
           height="12"
-          fill="#fff"
         />
       </clipPath>
     </defs>
@@ -42,14 +41,12 @@ const UsersMenuIcon = (props: SVGProps<SVGSVGElement>) => (
         data-name="Trazado 7088"
         d="M26.843,6.743a3.4,3.4,0,0,0,3.411-3.372,3.411,3.411,0,0,0-6.822,0,3.4,3.4,0,0,0,3.411,3.372"
         transform="translate(-22.334)"
-        fill="#fff"
       />
       <path
         id="Trazado_7089"
         data-name="Trazado 7089"
         d="M8.639,157.056a5.164,5.164,0,0,0-1.957-1.538,5.439,5.439,0,0,0-1.083-.362,5.2,5.2,0,0,0-1.117-.123c-.075,0-.151,0-.225.005H4.231a4.928,4.928,0,0,0-.549.059,5.236,5.236,0,0,0-3.276,1.92c-.029.039-.059.078-.086.116h0a1.723,1.723,0,0,0-.134,1.784,1.581,1.581,0,0,0,.255.356,1.559,1.559,0,0,0,.337.267,1.614,1.614,0,0,0,.4.167,1.743,1.743,0,0,0,.449.058H7.389a1.748,1.748,0,0,0,.452-.058,1.594,1.594,0,0,0,.4-.169,1.525,1.525,0,0,0,.335-.271,1.548,1.548,0,0,0,.251-.361,1.761,1.761,0,0,0-.191-1.85"
         transform="translate(0.001 -147.766)"
-        fill="#fff"
       />
     </g>
   </svg>
