+++ portal-ui/src/icons/TiersNotAvailableIcon.tsx
@@ -52,7 +52,6 @@ const TiersNotAvailableIcon = (props: SVGProps<SVGSVGElement>) => {
                   data-name="Trazado 441"
                   d="M13,3a.8.8,0,0,0-.392.092L4.374,7.482a.666.666,0,0,0,0,1.2l2.54,1.354-2.54,1.354a.666.666,0,0,0,0,1.2l2.54,1.353-2.54,1.354a.666.666,0,0,0,0,1.2l8.236,4.39a.8.8,0,0,0,.749,0l8.236-4.39a.666.666,0,0,0,0-1.2l-2.54-1.354,2.54-1.353a.666.666,0,0,0,0-1.2l-2.54-1.354L21.6,8.678a.666.666,0,0,0,0-1.2L13.36,3.092A.8.8,0,0,0,13,3ZM8.414,10.832l4.2,2.237a.8.8,0,0,0,.749,0l4.2-2.237,2.167,1.154-6.739,3.591L6.246,11.986Zm0,3.9,4.2,2.237a.8.8,0,0,0,.749,0l4.2-2.237,2.166,1.154-6.739,3.591L6.246,15.89Z"
                   transform="translate(-4 -3)"
-                  fill="#fff"
                 />
               </g>
             </g>
@@ -65,7 +64,6 @@ const TiersNotAvailableIcon = (props: SVGProps<SVGSVGElement>) => {
               cy="5"
               r="5"
               transform="translate(358 156)"
-              fill="#fff"
             />
             <path
               id="Elipse_593_-_Contorno"
